refactor: strike, ball 객체 분리
