+++ src/main/java/baseball/entity/Ball.java
@@ -1,0 +1,17 @@
+package baseball.entity;
+
+public class Ball {
+    private int count;
+
+    public int getCount() {
+        return count;
+    }
+
+    public void increaseCount() {
+        this.count++;
+    }
+
+    public void resetScore() {
+        this.count = 0;
+    }
+}
